[JBIDE-25670] deleting resources in an enforced sequence
Before: each resource was deleted in an individual thread resulting in
inappropriate removal ordering (ex. pod being deleted before rc, causing
pod tp be recreated etc.)
