+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/internal/core/job/AbstractDeleteResourceJob.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2015 Red Hat, Inc.
+ * Copyright (c) 2015-2018 Red Hat, Inc.
  * Distributed under license by Red Hat, Inc. All rights reserved.
  * This program is made available under the terms of the
  * Eclipse Public License v1.0 which accompanies this distribution,
@@ -8,7 +8,7 @@
  * Contributors:
  *     Red Hat, Inc. - initial API and implementation
  ******************************************************************************/
-package org.jboss.tools.openshift.internal.ui.job;
+package org.jboss.tools.openshift.internal.core.job;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
@@ -17,9 +17,8 @@ import org.eclipse.osgi.util.NLS;
 import org.jboss.tools.openshift.common.core.connection.ConnectionsRegistrySingleton;
 import org.jboss.tools.openshift.core.connection.Connection;
 import org.jboss.tools.openshift.core.connection.ConnectionProperties;
-import org.jboss.tools.openshift.core.connection.ConnectionsRegistryUtil;
 import org.jboss.tools.openshift.internal.common.core.job.AbstractDelegatingMonitorJob;
-import org.jboss.tools.openshift.internal.ui.OpenShiftUIActivator;
+import org.jboss.tools.openshift.internal.core.OpenShiftCoreActivator;
 
 import com.openshift.restclient.OpenShiftException;
 import com.openshift.restclient.model.IResource;
@@ -29,20 +28,19 @@ import com.openshift.restclient.model.IResource;
  * 
  * @author jeff.cantrill
  */
-public class DeleteResourceJob extends AbstractDelegatingMonitorJob {
+public abstract class AbstractDeleteResourceJob extends AbstractDelegatingMonitorJob {
 
-	private IResource resource;
-
-	public DeleteResourceJob(IResource resource) {
-		super("Delete Resource Job");
-		this.resource = resource;
+	public AbstractDeleteResourceJob(String name) {
+		super(name);
 	}
 
-	@Override
-	protected IStatus doRun(IProgressMonitor monitor) {
+	protected IStatus delete(final IResource resource, final Connection connection, final IProgressMonitor monitor) {
+		if (resource == null || monitor.isCanceled()) {
+			return Status.CANCEL_STATUS;
+		}
+
 		try {
-			monitor.beginTask("Delete Resource", IProgressMonitor.UNKNOWN);
-			Connection connection = ConnectionsRegistryUtil.getConnectionFor(resource);
+			monitor.subTask(NLS.bind("Delete Resource {0}...", resource.getName()));
 			if (connection != null) {
 				connection.deleteResource(resource);
 				ConnectionsRegistrySingleton.getInstance().fireConnectionChanged(connection,
@@ -50,10 +48,8 @@ public class DeleteResourceJob extends AbstractDelegatingMonitorJob {
 			}
 			return Status.OK_STATUS;
 		} catch (OpenShiftException e) {
-			return new Status(Status.ERROR, OpenShiftUIActivator.PLUGIN_ID,
+			return new Status(Status.ERROR, OpenShiftCoreActivator.PLUGIN_ID,
 					NLS.bind("Error deleting {0} named  {1}.", resource.getKind(), resource.getName()), e);
-		} finally {
-			monitor.done();
 		}
 	}
